Fix potiential null pointer dereference when initializing pointer fb
There is a sanity check at the end of evdi_painter_mode_changed_notify
to see if pointer 'painter' is null. This implies it is null at the
start of the function, in which case the assignment of fb using
fb = &painter->scanout_fb->base will be performing a null pointer
dereference.  Fix this by sanity checking if pointer is null to start
with and bail out early if it is.

Addresses-Coverity: ("Dereference before null check")
Signed-off-by: Colin Ian King <colin.king@canonical.com>