Removing workaround for Intel IOMMU domain allocation failure.
This was based on fix
https://patchwork.kernel.org/project/platform-driver-x86/patch/20190123230131.42094-1-sfruhwald@google.com/ that never get to kernel.
and merged in as PR https://github.com/DisplayLink/evdi/pull/161

Now that code is triggering crash on evdi module removal for newer kernel >= 5.15
Since this fix is doubious and we have no local reproduction I am removing it for newer kernels.
If this still happens on newer kernels we need to find better fix.
Or please use commit:
 Fix crash in iommu code on evdi module unload
   SHA: 26a6cbf9402d024a8497d4005a1ecab4c481f925

Tested on DELL Latitude 5491 with Intel Vitualization enabled and intel_iommu=on in boot options.